pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe-print"
version = "0.1.0"
description = "MCP server for end-to-end 3D printing automation with FDM printers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Hanna", email = "hunter.stradley@gmail.com"}
]
keywords = ["mcp", "3d-printing", "automation", "claude", "fdm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "numpy-stl>=3.0.0",
    "trimesh>=4.0.0",
    "paho-mqtt>=2.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

# CAD generation (CadQuery for parametric models)
cad = [
    "cadquery-ocp>=7.7.0",
]

# AI 3D generation (requires API keys)
ai = [
    "httpx>=0.25.0",
]

# All features
all = [
    "cadquery-ocp>=7.7.0",
]

[project.scripts]
vibe-print = "vibe_print:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vibe_print*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
